fix: minichef things (#313)
* feat: removed minichef data from redux

* fix: onchain minichef apr

* fix: only show farms with multiplyer greater than 0

* fix: multipler in on chain minichef data

the multiplyer cannot be divided by 100 because some multipliers are
smaller than 100 and when we do that it rounds the multipler to 0

* fix: graph query

* fix: graphql query

by default it only get the first 100 farms but if we have more this
it won't return all the farms

* feat: removed ust

* feat: added some coingecko currency id

* fix: removed unsed pool map

* fix: removed stakinginfo

can we replace this type by the new one minicheftstakinginfo
version 1: use the doublesidestakinginfo, the synthetix staking rewards contract
version 2: use the minicheftstakinginfo, the minichefv2
version 3: use pangochefinfo, the pangochef

* fix: usegetrewardtokens don't do unnecessary queries

* feat: added extra rewards tokens multipliers

* fix: tsc

* fix: minichef hooks folder
