+++ src/components/Pools/DetailModal/Details/index.tsx
@@ -9,13 +9,13 @@ import { ANALYTICS_PAGE } from 'src/constants';
 import { usePair } from 'src/data/Reserves';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
 import { convertCoingeckoTokens } from 'src/state/pcoingecko/hooks';
-import { useGetPoolDollerWorth } from 'src/state/pstake/hooks';
-import { StakingInfo } from 'src/state/pstake/types';
+import { useGetPoolDollerWorth } from 'src/state/pstake/hooks/common';
+import { DoubleSideStakingInfo } from 'src/state/pstake/types';
 import { unwrappedToken } from 'src/utils/wrappedCurrency';
 import { DetailsContainer } from './styled';
 
 type Props = {
-  stakingInfo: StakingInfo;
+  stakingInfo: DoubleSideStakingInfo;
 };
 
 const Details: React.FC<Props> = ({ stakingInfo }) => {
