+++ src/components/Pools/PoolCard/PoolCardV1.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { useTokens } from 'src/hooks/tokens/evm';
-import { StakingInfo } from 'src/state/pstake/types';
+import { DoubleSideStakingInfo } from 'src/state/pstake/types';
 import PoolCardView from './PoolCardView';
 
 export interface PoolCardV1Props {
-  stakingInfo: StakingInfo;
+  stakingInfo: DoubleSideStakingInfo;
   onClickViewDetail: () => void;
   version: number;
 }
@@ -14,7 +14,7 @@ const PoolCardV1 = ({ stakingInfo, onClickViewDetail, version }: PoolCardV1Props
 
   return (
     <PoolCardView
-      combinedApr={stakingInfo?.combinedApr}
+      combinedApr={stakingInfo.combinedApr}
       earnedAmount={stakingInfo?.earnedAmount}
       rewardTokens={rewardTokens}
       stakingInfo={stakingInfo}
