+++ src/components/Pools/PoolCard/PoolCardV3.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { PangoChefInfo } from 'src/state/ppangoChef/types';
-import { useGetRewardTokens } from 'src/state/pstake/hooks';
+import { useGetRewardTokens } from 'src/state/pstake/hooks/common';
 import PoolCardViewV3 from './PoolCardViewV3';
 
 export interface PoolCardProps {
@@ -10,7 +10,7 @@ export interface PoolCardProps {
 }
 
 const PoolCardV3 = ({ stakingInfo, onClickViewDetail, version }: PoolCardProps) => {
-  const rewardTokens = useGetRewardTokens(stakingInfo?.rewardTokens, stakingInfo?.rewardTokensAddress);
+  const rewardTokens = useGetRewardTokens(stakingInfo);
 
   return (
     <PoolCardViewV3
