+++ src/components/Pools/RemoveFarm/index.tsx
@@ -8,15 +8,15 @@ import { useChainId, usePangolinWeb3 } from 'src/hooks';
 import { MixPanelEvents, useMixpanel } from 'src/hooks/mixpanel';
 import { useGetHederaTokenNotAssociated, useHederaTokenAssociated } from 'src/hooks/tokens/hedera';
 import { usePangoChefWithdrawCallbackHook } from 'src/state/ppangoChef/hooks';
-import { useGetEarnedAmount, useGetRewardTokens, useMinichefPendingRewards } from 'src/state/pstake/hooks';
-import { StakingInfo } from 'src/state/pstake/types';
+import { useGetRewardTokens, useMinichefPendingRewards } from 'src/state/pstake/hooks/common';
+import { DoubleSideStakingInfo, MinichefStakingInfo } from 'src/state/pstake/types';
 import { useHederaPGLToken } from 'src/state/pwallet/hooks/hedera';
 import { hederaFn } from 'src/utils/hedera';
 import RemoveLiquidityDrawer from '../RemoveLiquidityDrawer';
 import { Buttons, FarmRemoveWrapper, RewardWrapper, Root, StatWrapper } from './styleds';
 
 interface RemoveFarmProps {
-  stakingInfo: StakingInfo;
+  stakingInfo: DoubleSideStakingInfo;
   version: number;
   onClose: () => void;
   // this prop will be used if user move away from first step
@@ -41,11 +41,11 @@ const RemoveFarm = ({ stakingInfo, version, onClose, onLoading, onComplete, redi
 
   const png = PNG[chainId];
 
-  const { rewardTokensAmount } = useMinichefPendingRewards(stakingInfo);
-  const rewardTokens = useGetRewardTokens(stakingInfo?.rewardTokens, stakingInfo?.rewardTokensAddress);
-  const isSuperFarm = (rewardTokensAmount || [])?.length > 0;
-
   const chefType = CHAINS[chainId].contracts?.mini_chef?.type ?? ChefType.MINI_CHEF_V2;
+
+  const { rewardTokensAmount } = useMinichefPendingRewards(stakingInfo);
+  const rewardTokens = useGetRewardTokens(stakingInfo);
+  const isSuperFarm = (rewardTokensAmount || [])?.length > 0;
 
   const mixpanel = useMixpanel();
 
@@ -81,7 +81,7 @@ const RemoveFarm = ({ stakingInfo, version, onClose, onLoading, onComplete, redi
 
   const { callback: withdrawCallback, error: withdrawCallbackError } = useWithdrawCallback({
     version,
-    poolId: stakingInfo?.pid,
+    poolId: chefType === ChefType.MINI_CHEF ? undefined : (stakingInfo as MinichefStakingInfo)?.pid,
     stakedAmount: stakingInfo?.stakedAmount,
     stakingRewardAddress: stakingInfo?.stakingRewardAddress,
   });
@@ -142,14 +142,10 @@ const RemoveFarm = ({ stakingInfo, version, onClose, onLoading, onComplete, redi
     error = withdrawCallbackError;
   }
 
-  const { earnedAmount } = useGetEarnedAmount(stakingInfo?.pid as string);
-
-  const newEarnedAmount = version !== 2 ? stakingInfo?.earnedAmount : earnedAmount;
+  const { earnedAmount } = stakingInfo;
 
   const token0 = stakingInfo.tokens[0];
   const token1 = stakingInfo.tokens[1];
-
-  const cheftType = CHAINS[chainId].contracts?.mini_chef?.type ?? ChefType.MINI_CHEF_V2;
 
   const renderButton = () => {
     if (!isHederaTokenAssociated && notAssociateTokens?.length > 0) {
@@ -190,11 +186,11 @@ const RemoveFarm = ({ stakingInfo, version, onClose, onLoading, onComplete, redi
                       />
                     </StatWrapper>
                   )}
-                  {newEarnedAmount && (
+                  {earnedAmount && (
                     <StatWrapper>
                       <Stat
                         title={t('earn.unclaimedReward', { symbol: png.symbol })}
-                        stat={newEarnedAmount?.toSignificant(4)}
+                        stat={earnedAmount?.toSignificant(4)}
                         titlePosition="top"
                         titleFontSize={12}
                         statFontSize={[20, 18]}
@@ -225,7 +221,7 @@ const RemoveFarm = ({ stakingInfo, version, onClose, onLoading, onComplete, redi
                 <Button
                   variant="primary"
                   onClick={
-                    cheftType === ChefType.PANGO_CHEF && !confirmRemove ? () => setConfirmRemove(true) : onWithdraw
+                    chefType === ChefType.PANGO_CHEF && !confirmRemove ? () => setConfirmRemove(true) : onWithdraw
                   }
                 >
                   {error ?? t('earn.withdrawAndClaim')}
