+++ src/constants/coingecko.ts
@@ -28,21 +28,21 @@ export const COINGECKO_CURRENCY_ID: { [chainId in ChainId]: string | undefined }
   [ChainId.NEAR_MAINNET]: 'near',
   [ChainId.NEAR_TESTNET]: undefined,
   [ChainId.COSTON2]: undefined,
-  [ChainId.ETHEREUM]: undefined,
-  [ChainId.POLYGON]: undefined,
-  [ChainId.FANTOM]: undefined,
-  [ChainId.XDAI]: undefined,
-  [ChainId.BSC]: undefined,
-  [ChainId.ARBITRUM]: undefined,
-  [ChainId.CELO]: undefined,
+  [ChainId.ETHEREUM]: 'ethereum',
+  [ChainId.POLYGON]: 'matic-network',
+  [ChainId.FANTOM]: 'fantom',
+  [ChainId.XDAI]: 'gnosis',
+  [ChainId.BSC]: 'binancecoin',
+  [ChainId.ARBITRUM]: 'ethereum', // arbitrum use eth as default coin
+  [ChainId.CELO]: 'celo',
   [ChainId.OKXCHAIN]: undefined,
   [ChainId.VELAS]: undefined,
-  [ChainId.AURORA]: undefined,
-  [ChainId.CRONOS]: undefined,
-  [ChainId.FUSE]: undefined,
-  [ChainId.MOONRIVER]: undefined,
-  [ChainId.MOONBEAM]: undefined,
-  [ChainId.OP]: undefined,
+  [ChainId.AURORA]: 'aurora-near',
+  [ChainId.CRONOS]: 'crypto-com-chain',
+  [ChainId.FUSE]: 'fuse-network-token',
+  [ChainId.MOONRIVER]: 'moonriver',
+  [ChainId.MOONBEAM]: 'moonbeam',
+  [ChainId.OP]: 'ethereum', // optimism use eth as default coin
   [ChainId.EVMOS_TESTNET]: undefined,
   [ChainId.EVMOS_MAINNET]: 'evmos',
 };
