+++ src/state/ppangoChef/hooks/hedera.ts
@@ -18,7 +18,7 @@ import { useTokens } from 'src/hooks/tokens/evm';
 import { usePangoChefContract } from 'src/hooks/useContract';
 import { usePairsCurrencyPrice } from 'src/hooks/useCurrencyPrice';
 import { useCoinGeckoCurrencyPrice } from 'src/state/pcoingecko/hooks';
-import { getExtraTokensWeeklyRewardRate } from 'src/state/pstake/hooks';
+import { getExtraTokensWeeklyRewardRate } from 'src/state/pstake/utils';
 import { useTransactionAdder } from 'src/state/ptransactions/hooks';
 import { useHederaPGLTokenAddresses, useHederaPairContractEVMAddresses } from 'src/state/pwallet/hooks/hedera';
 import { decimalToFraction } from 'src/utils';
@@ -573,13 +573,7 @@ export function useGetPangoChefInfosViaSubgraph() {
         (memo, rewardToken) => {
           const tokenObj = rewardToken.token;
           const _address = getAddress(tokenObj.id);
-          const _token = new Token(
-            chainId,
-            getAddress(tokenObj.id),
-            Number(tokenObj.decimals),
-            tokenObj.symbol,
-            tokenObj.name,
-          );
+          const _token = new Token(chainId, _address, Number(tokenObj.decimals), tokenObj.symbol, tokenObj.name);
           const _multiplier = JSBI.BigInt(rewardToken?.multiplier.toString());
 
           // remove png from rewards
