+++ src/state/psarstake/utils.ts
@@ -14,7 +14,7 @@ import { useSarStakingContract } from 'src/hooks/useContract';
 import { useUSDCPriceHook } from 'src/hooks/useUSDCPrice';
 import { maxAmountSpend } from 'src/utils/maxAmountSpend';
 import { CallState } from '../pmulticall/hooks';
-import { useDerivedStakeInfo } from '../pstake/hooks';
+import { useDerivedStakeInfo } from '../pstake/hooks/common';
 import { tryParseAmount } from '../pswap/hooks/common';
 import { useTransactionAdder } from '../ptransactions/hooks';
 import { useTokenBalanceHook } from '../pwallet/hooks';
