--- src/state/pstake/reducer.ts
@@ -1,86 +1,0 @@
-import { ChainId } from '@pangolindex/sdk';
-import { createReducer } from '@reduxjs/toolkit';
-import {
-  updateMinichefStakingAllAprs,
-  updateMinichefStakingAllData,
-  updateMinichefStakingAllFarmsEarnedAmount,
-} from './actions';
-import { MinichefV2 } from './types';
-
-export interface Apr {
-  pid: string;
-  swapFeeApr: number;
-  stakingApr: number;
-  combinedApr: number;
-}
-
-type MinichefStakingData = { [chainId in ChainId]: MinichefV2 };
-type Aprs = { [chainId in ChainId]: { [key: string]: Apr } };
-type EarnerdAmounts = { [chainId in ChainId]: { [key: string]: { pid: string; earnedAmount: number } } };
-
-export interface MinichefStakingInfoState {
-  readonly minichefStakingData: MinichefStakingData;
-  readonly aprs: Aprs;
-  readonly earnedAmounts: EarnerdAmounts;
-}
-
-const initialState: MinichefStakingInfoState = {
-  minichefStakingData: {} as MinichefStakingData,
-  aprs: {} as Aprs,
-  earnedAmounts: {} as EarnerdAmounts,
-};
-
-export enum SortingType {
-  totalStakedInUsd = 'totalStakedInUsd',
-  totalApr = 'totalApr',
-}
-
-export default createReducer(initialState, (builder) =>
-  builder
-    .addCase(
-      updateMinichefStakingAllData,
-      (
-        state,
-        {
-          payload: {
-            data: { chainId, data },
-          },
-        },
-      ) => {
-        // console.info('updateMinichefStakingAllData')
-
-        const existingData = { ...(state.minichefStakingData[chainId] || {}), ...data };
-        state.minichefStakingData[chainId] = existingData;
-      },
-    )
-
-    .addCase(
-      updateMinichefStakingAllAprs,
-      (
-        state,
-        {
-          payload: {
-            data: { chainId, data },
-          },
-        },
-      ) => {
-        // console.info('updateMinichefStakingALLAprs')
-        state.aprs[chainId] = data;
-      },
-    )
-
-    .addCase(
-      updateMinichefStakingAllFarmsEarnedAmount,
-      (
-        state,
-        {
-          payload: {
-            data: { chainId, data },
-          },
-        },
-      ) => {
-        // console.info('updateMinichefStakingAllFarmsEarnedAmount')
-        state.earnedAmounts[chainId] = data;
-      },
-    ),
-);
